+++ src/Raven.Studio/typescript/components/pages/database/aiHub/aiAgents/edit/utils/editAiAgentValidation.ts
@@ -66,6 +66,7 @@ const editSchema = yup.object({
             canRegenerateSchema: yup.boolean(),
             isSaved: yup.boolean(),
             isEditing: yup.boolean(),
+            prevValue: yup.object().nullable(),
         })
     ),
     actions: yup.array().of(
@@ -95,6 +96,7 @@ const editSchema = yup.object({
             canRegenerateSchema: yup.boolean(),
             isSaved: yup.boolean(),
             isEditing: yup.boolean(),
+            prevValue: yup.object().nullable(),
         })
     ),
